+++ Order.cs
@@ -6,7 +6,7 @@
         public DateTime Time { get; set; }
         public string Type { get; set; }
         public string Kind { get; set; }
-        public double Amount { get; set; }
-        public double Price { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Price { get; set; }
     }
 }
